"community edition" is not always a bad thing.
One of the projects I'd like to bring here is called "[Project] Community Edition".  It is completely open source and always has been, and is not an "up-sell" or "upgrade to pro" project, and there is no non-community-edition version. It is called community edition because it was taken over by the community after the company that used to maintain it stepped back (but it was always fully open source).

I think that `"up-sell" or "upgrade-to-pro"` covers what you are saying without unintentionally blocking legitimate community projects.

Thanks for considering.